+++ backend/movie-recommendation-backend/src/main/java/com/example/MovieRecommendationBackend/entity/User.java
@@ -8,7 +8,7 @@ public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "user_id")
-    private Long userId;
+    private Integer userId;
 
     @Column(name = "user_name")
     private String userName;
@@ -19,17 +19,17 @@ public class User {
     public User() {
     }
 
-    public User(Long userId, String userName, String userPassword) {
+    public User(Integer userId, String userName, String userPassword) {
         this.userId = userId;
         this.userName = userName;
         this.userPassword = userPassword;
     }
 
-    public Long getUserId() {
+    public Integer getUserId() {
         return userId;
     }
 
-    public void setUserId(Long userId) {
+    public void setUserId(Integer userId) {
         this.userId = userId;
     }
